+++ enclave_wrangler/dataset_download_new_api.py
@@ -3,24 +3,14 @@
 """Command Line Interface."""
 from argparse import ArgumentParser
 from typing import Dict
+from typeguard import typechecked
 
 import requests
-from typing import Dict
-from typeguard import typechecked
-import json
-import os
-from datetime import datetime, timezone
-import requests
-import pandas as pd
-import tempfile
 # import pyarrow as pa
-import pyarrow.parquet as pq
 # import asyncio
-import shutil
-import time
 
 from enclave_wrangler.config import config
-from enclave_wrangler.utils import log_debug_info
+# from enclave_wrangler.utils import log_debug_info
 
 HEADERS = {
     "authorization": f"Bearer {config['PALANTIR_ENCLAVE_AUTHENTICATION_BEARER_TOKEN']}",
@@ -29,20 +19,6 @@ HEADERS = {
 DEBUG = False
 # TARGET_CSV_DIR='data/datasets/'
 
-def ontocall(path) -> [{}]:
-    """API documentation at
-    https://www.palantir.com/docs/foundry/api/ontology-resources/object-types/list-object-types/
-    """
-
-    ontologyRid = config['ONTOLOGY_RID']
-    api_path = f'/api/v1/ontologies/{ontologyRid}/path'
-    url = f'https://{config["HOSTNAME"]}{api_path}'
-
-    response = requests.get(url, headers=HEADERS,)
-    response_json = response.json()
-    # print(response_json)
-    # types = pd.DataFrame(data=response_json)
-    return response_json['data']
 
 @typechecked
 def objTypes() -> [{}]:
@@ -60,10 +36,11 @@ def objTypes() -> [{}]:
     # types = pd.DataFrame(data=response_json)
     return response_json['data']
 
+
 def run() -> None:
+    """Run"""
     types = objTypes()
     print('\n'.join([t['apiName'] for t in types]))
-    pass
