+++ termhub/templates/errors/404.html
@@ -2,5 +2,5 @@
 {% block content %}
   <h1>Sorry ...</h1>
   <p>There's nothing here!</p>
-  <p><a href="{{url_for('index')}}">Back</a></p>
+  <p><a href="url_for('index')">Back</a></p>
 {% endblock %}
