+++ termhub/templates/errors/500.html
@@ -2,5 +2,5 @@
 {% block content %}
   <h1>Something's wrong!</h1>
   <p>Fortunately we are on the job, and you can just return home!</p>
-  <p><a href="{{url_for('index')}}">Back</a></p>
+<!--  <p><a href="url_for('index')">Back</a></p>-->
 {% endblock %}
