+++ termhub/templates/pages/browse_onto.html
@@ -2,28 +2,79 @@
 {% block title %}Termhub - Browse ontology{% endblock %}
 {% block content %}
     <script>
-        async function testjs(x) {
+      async function call_enclave_ontologies(path) {
+        p = fetch('/ontocall?path=' + path)
+        return await p.then(response => response.json())
+      }
+      async function get_enclave_objects(object_type) {
+        // TODO: temp get below to work
+        // Example response is a list of dictionairies that lok like this:
+        //  {
+        //    "properties": {
+        //      "conceptSetId": " Heavy menstrual bleeding",
+        //      "assignedInformatician": "4b054b72-ee25-48ca-9183-696cb9bff7ee",
+        //      "archived": true,
+        //      "stage": "Awaiting Editing",
+        //      "alias": "heavy_menstrual_bleeding",
+        //      "createdAt": "2021-07-28T19:20:04.268Z",
+        //      "createdBy": "976125fc-a13b-4571-a5bd-52c1918ff99b",
+        //      "conceptSetName": " Heavy menstrual bleeding",
+        //      "status": "Under Construction",
+        //      "intention": "Mixed"
+        //    },
+        objects = await call_enclave_ontologies('objects/' + object_type)
+        console.log(objects)
 
-            console.log(x)
-            p = fetch('/hello')
-            js = await p.then(response => response.json())
-
-            console.log(js)
-            //alert('hello world')
-        }
+        // https://codereview.stackexchange.com/questions/277042/generate-an-html-table-using-javascript-from-an-array-of-objects
+        // const notes = [{note: "n1", subject: "subject1", value: 10 }, {note: "n2", subject: "subject2", value: 15 }, {note: "n2", subject: "subject2", value: 5 }, {note: "n3", subject: "subject2", value: 20 }];
+        // // TODO: wanted below to work but .classList.add('table') said was void function so failed
+        // //  ...i could look up by ID and add, but i can't add the ID either becasue that's probably also a void func
+        // //  ...really wanting react
+        // // const tableTag = (tagName, props = {}) => Object.assign(document.createElement(tagName).classList.add('table'), props);
+        // const tag = (tagName, props = {}) => Object.assign(document.createElement(tagName), props);
+        // const txtTag = (tagName, str, props = {}) => tag(tagName, {textContent: str, ...props});
+        // const append = (el, ...sibs) => sibs.reduce((p, sib) => (p.appendChild(sib), p), el);
+        //
+        // append(document.body,
+        //   createTable("Subjects", "note", "subject", "value", notes),
+        //   createTable("Notes", "subject", "note", "value", notes),
+        //   createTable("Notes", "value", "note", "value", notes),
+        //   createTable("Values", "subject", "value", "value", notes));
+        //
+        // function createTable(title, colKey, rowKey, valKey, data) {
+        //   const byKey = key => [...new Set(data.map(rec => rec[key]))];
+        //   const byRowCol = (row, col) => data.reduce((val, rec) =>
+        //     val + (rec[rowKey] === row && rec[colKey] === col ? rec[valKey] : 0), 0);
+        //   const rows = byKey(rowKey), cols = byKey(colKey);
+        //   // TODO: How to add class="table" to table? etc. https://getbootstrap.com/docs/4.0/content/tables/
+        //   return append(tag("table", {"class": "table"}),
+        //     append(tag("tbody"),
+        //       append(tag("tr"),
+        //         txtTag("th", title),
+        //         ...cols.map(str => txtTag("th", str))
+        //       ),
+        //       ...rows.map(row =>
+        //         append(tag("tr"),
+        //           txtTag("th", row),
+        //           ...cols.map(col => txtTag("td", byRowCol(row, col)))
+        //         )
+        //       )
+        //     )
+        //   )
+        // }
+      }
     </script>
 
     <div style="float: left; padding: 10px; border: solid 1px brown;">
         <h1>Object types</h1>
         <ul>
           {% for object_type in object_types -%}
-            <!-- <li><a href="/concept-sets/{{object_type}}">{{ object_type }}</a></li -->
-            <li><a href="javascript:testjs('{{object_type}}')">{{ object_type }}</a></li>
+            <li><a href="javascript:get_enclave_objects('{{object_type}}')">{{ object_type }}</a></li>
           {% endfor %}
         </ul>
     </div>
 
-    <div style="float: left; width: 300px; height: 400px; border: solid 1px brown; margin-left: 30px">
+    <div id="populate" style="float: left; width: 300px; height: 400px; border: solid 1px brown; margin-left: 30px">
         <p>
             Hello!
         </p>
